move hasContext and hasValue to node file
